changed tree definition of texture to a graph one
previously, we had the texture defined as a tree,
i guess because thats what the attribute is called,
but really it should be a graph based on the way
connections can happen for complex textures. I also
developed more of the moon surface texture. it would be
nice to be able to automatically space out the nodes when they're placed.
maybe we can come up with a simple method of spacing them out, even if it is random