texture import from blender
I started to make a function for importing textures made in blender.
it occured to me that making the textures in blender is probably
easier than making them manually with a text file,
but saving them in a text file for reuse is still preferred.
making the function is going to be tricky because recording all the properties will require some special programming